Update Package.swift (#9)
### Motivation

The example points to a local version of the repo, but it's now ready to
point to github.

### Modifications

Changed the `Package.swift` in the getting started example to point to
this repo.

### Result

Example code will no longer look for a local copy. 

### Test Plan

I built and ran the package and it said hello!