Default to enabled observability (#54)
### Motivation

Swift Log/Swift Metrics integration should be enabled by default. And
the operator can always pass in a no-op log/metrics backend if they want
to discard it.

This is important when Configuration is used as a transitive dependency,
as there the operator wouldn't have a way to _enable_ observability from
this library unless the direct dependency that pulled in Configuration
was cooperative.

### Modifications

Enabled logging/metrics by default by changing the default parameter
values in the reloading providers.

### Result

Logs/metrics emitted for reloading providers by default.

### Test Plan

Adapted tests.