NonisolatedNonsendingByDefault and other concurrency fixes (#63)
### Motivation

- Fixes https://github.com/apple/swift-configuration/issues/17
- Fixes https://github.com/apple/swift-configuration/issues/23
- Fixes https://github.com/apple/swift-configuration/issues/32
- Fixes https://github.com/apple/swift-configuration/issues/43
- Fixes https://github.com/apple/swift-configuration/issues/44

These 5 fixes are intertwined and were difficult to split out, so here's
one PR that fixes them all.

### Modifications

- The PR enables NonisolatedNonsendingByDefault, which caused an error
in the existing combineLatest implementation, which we wanted to replace
anyway.
- So we vendor in the code from Swift Async Algoritms (the PR is open
here: https://github.com/apple/swift-async-algorithms/pull/360), the
plan is to remove this copy when it lands in async algos.
- Updating the combineLatest implementation to the correct one requires
the async sequences to be Sendable, which in turn requires Swift 6.2 for
us to be able to spell `any (AsyncSequence & Sendable)` correctly, it
doesn't work on 6.1.
- We also enabled the explicit sendable warning in CI, helping us ensure
all our public API is explicitly annotated as Sendable or not.

### Result

Addressed 5 important issues that impact the API and concurrency.

### Test Plan

Brought over tests for the new combineLatest implementation, added more
unit tests for failure cases when using the ConfigReader with multiple
providers.