Allow non-copyable Return types in with-style methods (#81)
### Motivation

Fixes https://github.com/apple/swift-configuration/issues/77.

Users might create a value of a non-copyable type and return it from one
of the watch methods. Today, that's not possible, because the `Return`
type implicitly requires `Copyable`. But that's just because it's the
default, we don't do any copying in the watch methods themselves, so we
can relax it.

### Modifications

Allow returning non-copyable values from watch methods.

### Result

Easier usage of non-copyable types with Swift Configuration.

### Test Plan

Adapted tests, all pass.