Add support for one zero and yes no boolean strings in environement variables provider (#120)
### Motivation

Add support for 1, 0, and YES, NO boolean strings in
EnvironmentVariablesProvider. Support for YES and NO boolean strings was
not discussed in the original issue. From personal experience, I found
it interesting to include it. If rejected, I will remove it. I tried to
remain consitent with the existing decoders and the way they are used by
the provider.

Closes #110 

### Modifications

The EnvironmentVariablesProvider's bool decoder is configurable and can
be initialized with support for true/false, 1/0, yes/no, or any
combination of them.

### Result

Usage of EnvironmentVariablesProvider has not changed, but by default
all pairs of boolean strings are accepted when retrieving values as Bool
or BoolArray. Client code can configure the provider to use a bool
decoder with support only for true/false strings.

If it is more important to keep the behavior as it was, meaning default
support for true/false only in the provider, then I will modify the
default set for the EnvironmentVariablesProvider BoolDecoder.

### Test Plan

The decoder is tested with dedicated tests, and additional tests for
receiving a value from a provider were added to test the integration of
BoolDecoder with EnvironmentVariablesProvider.

Code is commented, and tests are green locally.

---------

Co-authored-by: Honza Dvorsky <honza@apple.com>